fix(build.zig): emcscripten sysroot absolute path
When compiling with Zig for emscripten the `--sysroot` argument must be
given.

A LazyPath, when constructed with `b.path`, must be given relative
paths. Therefore, if `--sysroot` is given an absolute path, an error
such as

    thread 637865 panic: sub_path is expected to be relative to the
    build root, but was this absolute path:
    '/home/user/emsdk/upstream/emscripten/cache/sysroot/include'.
    It is best avoid absolute paths, but if you must, it is supported by
    LazyPath.cwd_relative

occurs. Using a relative path instead results in

    /home/user/zig/lib/std/fs.zig:280:11: 0x10f0b3a in openDirAbsolute
    (build) assert(path.isAbsolute(absolute_path));

where an assertion that the path that is being opened is absolute. This
commit mandates that the path `--sysroot` path is absolute by using a
LazyPath without a `root`, i.e. created without using `b.path`.

Works with Zig 0.12.0.